# swagger.yml
basePath: /
consumes:
- application/json
- application/xml
host: 127.0.0.1:6420
info:
  contact:
    email: skycoin.doe@example.com
    name: Skycoin project
    url: http://127.0.0.1:6420
  description: Skycoin is a next-generation cryptocurrency.
  title: Skycoin REST API.
  version: 0.25.1
paths:
  /api/v1/address_uxouts:
    get:
      description: Returns the historical, spent outputs associated with an address
      operationId: addressUxouts
      parameters:
      - description: address to filter by
        in: query
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Response for endpoint /api/v1/address_uxouts
          schema:
            items:
              properties:
                coins:
                  format: integer
                  type: integer
                hours:
                  format: int64
                  type: integer
                owner_address:
                  type: string
                spent_block_seq:
                  format: in64
                  type: integer
                spent_tx:
                  type: string
                src_block_seq:
                  format: int64
                  type: integer
                src_tx:
                  type: string
                time:
                  format: int64
                  type: integer
                uxid:
                  type: string
            type: array
        default:
          $ref: '#/responses/genericError'
  /api/v1/addresscount:
    get:
      operationId: addressCount
      produces:
      - application/json
      responses:
        "200":
          description: This endpoint Returns the total number of unique address that
            have coins.
          schema:
            properties:
              count:
                format: int64
                type: integer
        default:
          $ref: '#/responses/genericError'
      summary: Returns the total number of unique address that have coins.
  /api/v1/balance:
    get:
      operationId: balanceGet
      parameters:
      - description: command separated list of addresses
        in: query
        name: addrs
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns the balance of one or more addresses
          schema:
            properties:
              addresses:
                additionalProperties:
                  properties:
                    coins:
                      format: int64
                      type: integer
                    hours:
                      format: int64
                      type: integer
                  type: object
                type: object
              confirmed:
                properties:
                  coins:
                    format: int64
                    type: integer
                  hours:
                    format: int64
                    type: integer
                type: object
              predicted:
                properties:
                  coins:
                    format: int64
                    type: integer
                  hours:
                    format: int64
                    type: integer
                type: object
        default:
          $ref: '#/responses/genericError'
      summary: Returns the balance of one or more addresses, both confirmed and predicted.
        The predicted balance is the confirmed balance minus the pending spends.
    post:
      operationId: balancePost
      parameters:
      - description: command separated list of addresses
        in: query
        name: addrs
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Returns the balance of one or more addresses
          schema:
            properties:
              addresses:
                additionalProperties:
                  properties:
                    coins:
                      format: int64
                      type: integer
                    hours:
                      format: int64
                      type: integer
                  type: object
                type: object
              confirmed:
                properties:
                  coins:
                    format: int64
                    type: integer
                  hours:
                    format: int64
                    type: integer
                type: object
              predicted:
                properties:
                  coins:
                    format: int64
                    type: integer
                  hours:
                    format: int64
                    type: integer
                type: object
        default:
          $ref: '#/responses/genericError'
      security:
      - csrfAuth: []
      summary: Returns the balance of one or more addresses, both confirmed and predicted.
        The predicted balance is the confirmed balance minus the pending spends.
  /api/v1/block:
    get:
      description: 'Returns a block by hash or seq. Note: only one of hash or seq
        is allowed'
      operationId: block
      parameters:
      - in: query
        name: hash
        type: string
      - in: query
        name: seq
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: BlockVerbose represents a readable block with verbose data.
          schema:
            properties:
              body:
                properties:
                  txns:
                    items:
                      properties:
                        inner_hash:
                          type: string
                        inputs:
                          items:
                            type: string
                          type: array
                        length:
                          format: int32
                          type: integer
                        outputs:
                          items:
                            properties:
                              coins:
                                type: string
                              dst:
                                type: string
                              hours:
                                format: int64
                                type: integer
                              uxid:
                                type: string
                          type: array
                        sigs:
                          items:
                            type: string
                          type: array
                        timestamp:
                          format: int64
                          type: integer
                        txid:
                          type: string
                        type:
                          format: int32
                          type: integer
                    type: array
                type: object
              header:
                properties:
                  block_hash:
                    type: string
                  fee:
                    format: int32
                    type: integer
                  previous_block_hash:
                    type: string
                  seq:
                    format: int32
                    type: integer
                  timestamp:
                    format: int64
                    type: integer
                  tx_body_hash:
                    type: string
                  ux_hash:
                    type: string
                  version:
                    format: int32
                    type: integer
                type: object
              size:
                format: int32
                type: integer
        default:
          $ref: '#/responses/genericError'
  /api/v1/block/verbose:
    get:
      description: |-
        Returns a block by hash or seq. Note: only one of hash or seq is allowed
        With verbose.
      operationId: blockVerbose
      parameters:
      - in: query
        name: hash
        type: string
      - in: query
        name: seq
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: BlockVerbose represents a readable block with verbose data.
          schema:
            properties:
              body:
                properties:
                  txns:
                    items:
                      properties:
                        fee:
                          format: int32
                          type: integer
                        hash:
                          type: string
                        inner_hash:
                          type: string
                        inputs:
                          items:
                            properties:
                              calculated_hours:
                                format: int64
                                type: integer
                              coins:
                                type: string
                              dst:
                                type: string
                              hours:
                                format: int64
                                type: integer
                              uxid:
                                type: string
                          type: array
                        length:
                          format: int32
                          type: integer
                        outputs:
                          items:
                            properties:
                              coins:
                                type: string
                              dst:
                                type: string
                              hours:
                                format: int64
                                type: integer
                              uxid:
                                type: string
                          type: array
                        sigs:
                          items:
                            type: string
                          type: array
                        type:
                          format: int32
                          type: integer
                    type: array
                type: object
              header:
                properties:
                  block_hash:
                    type: string
                  fee:
                    format: int32
                    type: integer
                  previous_block_hash:
                    type: string
                  seq:
                    format: int32
                    type: integer
                  timestamp:
                    format: int64
                    type: integer
                  tx_body_hash:
                    type: string
                  ux_hash:
                    type: string
                  version:
                    format: int32
                    type: integer
                type: object
              size:
                format: int32
                type: integer
        default:
          $ref: '#/responses/genericError'
  /api/v1/blockchain/metadata:
    get:
      operationId: blockchainMetadata
      produces:
      - application/json
      responses:
        "200":
          description: This endpoint returns the blockchain metadata.
          schema:
            properties:
              head:
                properties:
                  block_hash:
                    type: string
                  fee:
                    format: int64
                    type: integer
                  previous_block_hash:
                    type: string
                  seq:
                    type: string
                  timestamp:
                    format: int64
                    type: integer
                  tx_body_hash:
                    type: string
                  ux_hash:
                    type: string
                  version:
                    format: int64
                    type: integer
                type: object
              unconfirmed:
                format: int64
                type: integer
              unspents:
                format: int64
                type: integer
        default:
          $ref: '#/responses/genericError'
      summary: Returns the blockchain metadata.
  /api/v1/blockchain/progress:
    get:
      operationId: blockchainProgress
      produces:
      - application/json
      responses:
        "200":
          description: This endpoint returns the blockchain sync progress
          schema:
            properties:
              current:
                format: int64
                type: integer
              highest:
                format: int64
                type: integer
              peers:
                items:
                  properties:
                    address:
                      type: string
                    height:
                      format: int64
                      type: integer
                type: array
        default:
          $ref: '#/responses/genericError'
      summary: Returns the blockchain sync progress.
...
  /api/v1/coinSupply:
    get:
      description: coinSupplyHandler returns coin distribution supply stats
      operationId: coinSupply
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/coinSupply'
        default:
          $ref: '#/responses/genericError'
      schemes:
      - http
      - https
  /api/v1/csrf:
    get:
      operationId: csrf
      produces:
      - application/json
      responses:
        "200":
          description: Return a csrf Token.
          schema:
            properties:
              csrf_token:
                type: string
            type: object
        default:
          $ref: '#/responses/genericError'
      summary: Creates a new CSRF token. Previous CSRF tokens are invalidated by this
        call.
...
  /api/v1/network/connection/disconnect:
    post:
      description: This endpoint disconnects a connection by ID or address
      operationId: networkConnectionsDisconnect
      parameters:
      - description: Address id.
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: This endpoint
        default:
          $ref: '#/responses/genericError'
      security:
      - csrfAuth: []
...
  /api/v1/network/connections/exchange:
    get:
      description: This endpoint returns all connections found through peer exchange
      operationId: networkConnectionsExchange
      produces:
      - application/json
      responses:
        "200":
          description: This endpoint return a list of all connections found through
            peer exchange.
          schema:
            items:
              type: string
            type: array
        default:
          $ref: '#/responses/genericError'
...
produces:
- application/json
- application/xml
responses:
...
  genericError:
    description: |-
      A GenericError is the default error message that is generated.
      For certain status codes there are more appropriate error structures.
    schema:
      properties:
        code:
          format: int32
          type: integer
          x-go-name: Code
        message:
          type: string
          x-go-name: Message
      type: object
schemes:
- http
- https
securityDefinitions:
  csrfAuth:
    in: header
    name: X-CSRF-TOKEN
    type: apiKey
swagger: "2.0"